feat: Add support for linting @nuxtjs/sanity
The Nuxt community module for Sanity integration brings its own export
of `groq` to be a "batteries included" experience. While a plugin for
VS Code can bring syntax highlighting it would be nice to have the
option of using this existing ESLint plugin to lint GROQ in projects
using `@nuxtjs/sanity` without duplicating `groq`.

This commit adds some suggested changes to add support for this other
import. It adds a minimum required Node version in `engines`.